+++ scripts/pypsa-de/modify_prenetwork.py
@@ -829,83 +829,86 @@ def aladin_mobility_demand(n):
     """
     Change loads in Germany to use Aladin data for road demand.
     """
+    logger.info(
+        "Overwriting land transport demand with Aladin data. In particular the `land_transport_electric_share` config setting will not be used."
+    )
     # get aladin data
     aladin_demand = pd.read_csv(snakemake.input.aladin_demand, index_col=0)
 
     simulation_period_correction_factor = n.snapshot_weightings.objective.sum() / 8760
 
     # oil demand
-    oil_demand = aladin_demand.Liquids * simulation_period_correction_factor
-    oil_index = n.loads[
-        (n.loads.carrier == "land transport oil") & (n.loads.index.str[:2] == "DE")
-    ].index
-    oil_demand.index = [f"{i} land transport oil" for i in oil_demand.index]
+    if "land transport oil" in n.loads.carrier.unique():  # i.e. before 2050
+        oil_demand = pd.Series(
+            aladin_demand.Liquids.values * simulation_period_correction_factor,
+            index=aladin_demand.index + " land transport oil",
+        )
 
-    profile = n.loads_t.p_set.loc[:, oil_index]
-    profile /= profile.sum()
-    n.loads_t.p_set.loc[:, oil_index] = (oil_demand * profile).div(
-        n.snapshot_weightings.objective, axis=0
-    )
+        profile = n.loads_t.p_set.loc[:, oil_demand.index]
+        profile /= profile.sum()
+        n.loads_t.p_set.loc[:, oil_demand.index] = (oil_demand * profile).div(
+            n.snapshot_weightings.objective, axis=0
+        )
 
     # hydrogen demand
-    h2_demand = aladin_demand.Hydrogen * simulation_period_correction_factor
-    h2_index = n.loads[
-        (n.loads.carrier == "land transport fuel cell")
-        & (n.loads.index.str[:2] == "DE")
-    ].index
-    h2_demand.index = [f"{i} land transport fuel cell" for i in h2_demand.index]
+    h2_demand = pd.Series(
+        aladin_demand.Hydrogen.values * simulation_period_correction_factor,
+        index=aladin_demand.index + " land transport fuel cell",
+    )
 
-    profile = n.loads_t.p_set.loc[:, h2_index]
+    profile = n.loads_t.p_set.loc[:, h2_demand.index]
     profile /= profile.sum()
-    n.loads_t.p_set.loc[:, h2_index] = (h2_demand * profile).div(
+    n.loads_t.p_set.loc[:, h2_demand.index] = (h2_demand * profile).div(
         n.snapshot_weightings.objective, axis=0
     )
 
     # electricity demand
-    ev_demand = aladin_demand.Electricity * simulation_period_correction_factor
-    ev_index = n.loads[
-        (n.loads.carrier == "land transport EV") & (n.loads.index.str[:2] == "DE")
-    ].index
-    ev_demand.index = [f"{i} land transport EV" for i in ev_demand.index]
+    ev_demand = pd.Series(
+        aladin_demand.Electricity.values * simulation_period_correction_factor,
+        index=aladin_demand.index + " land transport EV",
+    )
 
-    profile = n.loads_t.p_set.loc[:, ev_index]
+    profile = n.loads_t.p_set.loc[:, ev_demand.index]
     profile /= profile.sum()
-    n.loads_t.p_set.loc[:, ev_index] = (ev_demand * profile).div(
+    n.loads_t.p_set.loc[:, ev_demand.index] = (ev_demand * profile).div(
         n.snapshot_weightings.objective, axis=0
     )
 
     # adjust BEV charger and V2G capacities
-    number_cars = pd.read_csv(snakemake.input.transport_data, index_col=0)[
-        "number cars"
-    ].filter(like="DE")
-
-    factor = (
-        aladin_demand.number_of_cars
-        * 1e6
-        / (
-            number_cars
-            * snakemake.params.land_transport_electric_share[
-                int(snakemake.wildcards.planning_horizons)
-            ]
-        )
-    )
 
     BEV_charger_i = n.links[
         (n.links.carrier == "BEV charger") & (n.links.bus0.str.startswith("DE"))
     ].index
-    n.links.loc[BEV_charger_i].p_nom *= pd.Series(factor.values, index=BEV_charger_i)
+
+    # Check that buses in network and aladin_data appear in same order
+    assert [
+        idx.startswith(idx2) for (idx, idx2) in zip(BEV_charger_i, aladin_demand.index)
+    ]
+
+    # Then directly use .values for assignment
+    p_nom = (
+        aladin_demand.number_of_cars.values * 1e6 * snakemake.params.bev_charge_rate
+    )  # same logic like in prepare_sector_network
+
+    n.links.loc[BEV_charger_i, "p_nom"] = p_nom
 
     V2G_i = n.links[
         (n.links.carrier == "V2G") & (n.links.bus0.str.startswith("DE"))
     ].index
     if not V2G_i.empty:
-        n.links.loc[V2G_i].p_nom *= pd.Series(factor.values, index=V2G_i)
+        n.links.loc[V2G_i, "p_nom"] = p_nom * snakemake.params.bev_dsm_availability
 
     dsm_i = n.stores[
         (n.stores.carrier == "EV battery") & (n.stores.bus.str.startswith("DE"))
     ].index
+    e_nom = (
+        aladin_demand.number_of_cars.values
+        * 1e6
+        * snakemake.params.bev_energy
+        * snakemake.params.bev_dsm_availability
+    )
     if not dsm_i.empty:
-        n.stores.loc[dsm_i].e_nom *= pd.Series(factor.values, index=dsm_i)
+        n.stores.loc[dsm_i, "e_nom"] = e_nom
 
 
 def add_hydrogen_turbines(n):